requirements.txt
py-cpuinfo
gitpython
moviepy
opencv-python
scikit-build
typing
diskcache